@@ -19,26 +19,25 @@
     },
     "devDependencies": {
         "@babel/core": "^7.12.10",
-        "@babel/preset-env": "^7.12.1",
+        "@babel/preset-env": "^7.12.11",
         "@babel/preset-react": "^7.12.10",
         "autoprefixer": "^9.7.4",
-        "babel-loader": "^8.1.0",
+        "babel-loader": "^8.2.2",
         "core-js": "^3.8.2",
         "css-loader": "^5.0.1",
         "mini-css-extract-plugin": "^1.3.3",
         "npm-run-all": "^4.1.5",
         "postcss-loader": "^4.1.0",
         "react": "16.14.0",
-        "react-dom": "16.8.6",
+        "react-dom": "16.14.0",
         "react-instantsearch-dom": "6.0.0",
         "react-scripts": "4.0.1",
         "tap-spot": "^1.1.1",
-        "webpack": "^4.42.0",
+        "webpack": "^4.46.0",
         "webpack-cli": "^4.2.0"
     },
     "dependencies": {
-        "@algolia/algoliasearch-netlify-frontend": "^1.0.3",
-        "@babel/cli": "^7.11.6",
+        "@babel/cli": "^7.12.10",
         "@munter/tap-render": "^0.2.0",
         "algoliasearch": "^4.8.3",
         "angular": "^1.8.1",
@@ -46,7 +45,7 @@
         "assets-webpack-plugin": "^6.1.2",
         "atomic-algolia": "^0.3.19",
         "axios": "^0.21.1",
-        "bootstrap": "^4.5.2",
+        "bootstrap": "^4.5.3",
         "chalk": "^4.1.0",
         "child_process": "^1.0.2",
         "clean-webpack-plugin": "^3.0.0",
@@ -55,10 +54,10 @@
         "glob": "^7.1.6",
         "globby": "^11.0.1",
         "highlight.js": "^10.5.0",
-        "html-entities": "^1.3.3",
-        "html-webpack-plugin": "^4.5.0",
+        "html-entities": "^2.0.2",
+        "html-webpack-plugin": "^4.5.1",
         "hyperlink": "^4.6.0",
-        "instantsearch.js": "^4.8.1",
+        "instantsearch.js": "^4.10.0",
         "jquery": "^3.5.1",
         "moment": "^2.29.1",
         "mustache": "^4.0.1",
